+++ data/src/main/java/com/nexters/boolti/data/di/RepositoryModule.kt
@@ -2,10 +2,8 @@ package com.nexters.boolti.data.di
 
 import com.nexters.boolti.data.repository.AuthRepositoryImpl
 import com.nexters.boolti.data.repository.ConfigRepositoryImpl
-import com.nexters.boolti.data.repository.SignUpRepositoryImpl
 import com.nexters.boolti.domain.repository.AuthRepository
 import com.nexters.boolti.domain.repository.ConfigRepository
-import com.nexters.boolti.domain.repository.SignUpRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -19,7 +17,4 @@ abstract class RepositoryModule {
 
     @Binds
     abstract fun bindAuthRepository(repository: AuthRepositoryImpl): AuthRepository
-
-    @Binds
-    abstract fun bindSignUpRepository(repository: SignUpRepositoryImpl): SignUpRepository
 }
