--- domain/src/main/java/com/nexters/boolti/domain/repository/SignUpRepository.kt
@@ -1,7 +1,0 @@
-package com.nexters.boolti.domain.repository
-
-import com.nexters.boolti.domain.request.SignUpRequest
-
-interface SignUpRepository {
-    suspend fun signUp(signUpRequest: SignUpRequest): Result<Unit>
-}
